2025/06/24 13:48 turn testing done
So hey its been actually not that long, I've been working on this for a couple of days and it is amazing. I am finished with the testing the next part will be the implementation to the core scripts and making sure everything else will click into places, but I feel like it will, and the harder part of the logic has been done. It is quite simple. First the player's turn will play out they can choose whatever they want and after the last player is done, than the turn for the enemies to choose will play out. The enemies get to sort who they want, for example all the players in front or behind them self, than they get to sort out all of the enemies and then finally, they get all the player's actions as a reward. From there we can just implement logic to do whatever we feel like! When every character's option has been submitted, the play out script will then do the rest and puts all the actions into motion.

It is cool to see working out, because I don't have to worry about enemy behavior until the 2nd part of the topology and player's can't really mess up that much of the code while it's their turn. It is simple and fun and I think this is my favorite thing I did in this project so far.

For now that is it, I will try to commit one more before todays end, but we will see if I can connect all of the topology into motion by 7 pm. I can just hope.