2025/06/24 18:09 HAHAHAHHAHA I was so wrong!
HAHAHAhaha remember when I written down in the last commit that I hope I will be done with it by 7 PM???

Well funny bit I just did there cause I will probably not going to be done with this until probably next week HAHAHA!

The meat and bones of the entire joke I'm just describing is the fact that it has been 2... NO! 3 months since I last commit I did and 4 month since I stopped editing my code, so you can imagen how god damn surprising it is when you look at a code you wrote and you are just petrified by it. No joke I was sitting in front of my monitor for a solid 10 minutes trying to understand what the absolute FUCK I was thinking when I wrote this terrible piece of shit. No joke I was trying to understand and follow back the trails of my 4 months ago self putted down lead and for around 2 and a half hours straight trying and failing to find some sort of lead where I can just put the pieces together and start to integrate my new code, but I either found a part of a code that has already been taken care of OR a back ally of a darkened street, with only more and more alleyways making a labyrinth of WHAT THE FUCK. I tried and yet I still came out not knowing everything. I'm writing this as of the moment of me giving up atm, because I just have no fucking clue where to even start. CAN YOU IMAGEN A MAZE WHERE YOU FEEL LIKE THERE ARE NO STARTS OR ENDS AND YOUR ONLY LEADS ARE SIGNS THAT FEELS LIKE POINTING TO EACH OTHER? I don't think so and neither do I, because it is physically impossible us to comprehend.

So I'm writing this as a fuck you to myself in the past, because not only I couldn't code better understandable code, but I couldn't even make my comments comprehensible. So thank you, and fuck you!

...
Anyways I will TRY my best to make that mess better... but for now I will go and take a fucking shower! Good night!