2025/07/06 17:15 backend changes
I've done some work on implementing the new turn handler, but there are still so much to be done. Any other news is that HOLY SHIT IT'S been 2 weeks AAAAAA. I really wanted to work on this more, but I got so busy and now I kinda just feel bad. BUT... progress is progress hehe...

I think that is all I will share, nothing more than work stuff, and private life stuff, but yeah, let's hope the next commit will be the full backend change...

OH and I remembered that I will be 100% not posting after this week, because of vacation... yay