2025/07/08 17:14 DONE WITH THE BACK END STUFF WOHOO
And that is done! With the final code change in place I'm finally back at the same knowledge as 3-4 month before and the new system is finally in place I can do so much more than before, I still need better documentation, reorganizing the Trello page, adding some new stuff such as enemy AI and other stuff I have to fix like for example I started working on action indicators which is something I desperately need because for all of this time I watched the turn counter from the 'Remote' menu and I'm having ENOUGH. Other than that I will implement the Enemy AI tomorrow if everything will go according to plan and that will be IT. For 2 straight weeks I will be code less... again because life is really funny... so... what else to say than probably this will be my second to last update on the project, after that I will try to fix some bugs and reorganize the Trello page to do what I like best... ❇️SORTING ❇️

But yeah that is about it... I can't think of anything else that would be important... hmm

maybe the one and only thing that I can consider as news is the fact that Mondocon is at this weekend so that is the reason for the 2 weeks of nothingness. And so you future self know this was where you dressed up as a wizard with a stick that you found in you mother's garden... not much to say about it. I have no idea what will happen in this Con, but I hope I don't need to explain or mention it again... hopefully...

Anyways BYEEEEE