2025/08/05 19:05 I'M BACK BITCHES!!!
oh wow it's been a month huh...

well those 2 weeks were a bit extended huh?

… Anyways I done some stuff I will start soon on adding defeat and KO actions to my game. The plan should be that I want to make a simple RPG system before I add back the moving part. For now I wish not to focus on the stupid and over complicated 2.5D problem even if it is relevant in some way. I been also experimenting with Animation Trees and Pixel Composer a bit, but not much.

As for my life well.... I came back from vacation and than my pc's GPU decided to not work anymore... no clue what could have caused it, but the good news is that I Have a new PC and a better stronger one too. It has an RTX   3000 in it and it runs smoother than my old one. Although it has its quirks, it isn't too bad to be annoying or outrages. And so I hope it will last until I graduate. Oh yeah I was admitted to the university.

YAAAY 🎊

Anyhow... At the moment I'm writing it is still unclear if I will be able to get into collage. Which makes me... depressed... lol
So yeah there are still things have to be done but I will say this... Much of the stuff for LOTTH is going on it's beautiful way. I feel like as soon I can get this RPG done the sooner I can start working on the game and the sooner I will be able to make a team who could help me out make this game.

For now I just want to focus on finishing a presentable game I can show of, test and finalize. Everything else feels less important and I have less and less things to do atm. But I have a clue that it will not gonna last forever so the plan is this:

finish the game before next year!

If and that is a HUGE IF. I could have a the prototype ready than I could just continue to build a team and slowly start to work on level design, story, character design, platformer section and much much more I still need to do! I know I haven't even kept the last promise I made so I will just going to make my mouth where my hand is and just do it if it don't work out than it will don't, but I do not wish to continue the PROTOTYPE into the next year, and instead start the actual game with everything. It will be... hard and long and painfully difficult, but I should try to make more commits and sit down and work as much as I can.

For today I will just sit back and take a break, but I will continue almost immediately as soon as I can tomorrow.