2025/08/09 18:42 starting to get annoyed :(
So what happened i that I forgot to take it into count that I need to also take out characters who fall both after the battle and in the middle of the turn. It is really my bad that I didn't accounted for it, but I just have no real idea how to approach this stage by stage, so I just thrown my head against the wall and hope it will bash through the other side. I should and probably will eventually need to redo this, because it seems way to complicated for something so simple in high sight. So what I will do is that I will just continue trying to finish it for the best of my ability. We will see, I will write again tomorrow (hopefully).