2025/08/12 14:11 Done with the freed characters handling
It was much simpler than I thought, but still gave me trouble that is for sure! The solution was really simple and I should have done it way sooner. Removed a bunch of the [node of character].[character group].[index]... and just it is the node instead of a pointer that tries to tell you which character is. Much easier and funner to handle.

And with that I'm partially done with KO and death, but I still need to add animation to enemy ballers and would love to bug find a bit. So the plan is that after this I will look into the enemy AI which I just briefly touched on. Perhaps create an actual handler this time god damn it.

And after that I will hope to finally tackle the big fish again... the ellipsis trouble!
(I can't even type ellipsis anymore XD)
But yeah that should be a plan I like, it should be fine for the time being.
That is for it today see ya ;P