2025/08/25 13:46 failure
the same bug is still there, I don't think I can continue without fixing that bug. It is just SOOO annoying because for some reason it just duplicates on both enemy for seemingly no reason. I have little clue on why, and I just cannot figure it out. I will need to redo it. And the core reason for the idea of redoing it is, that in the testing area it works how it should, one of them has X data doing that and the other has lets say Y data increased, but with my case one of them receives the data and the next guy just takes the previous guy data and continues... I do not understand why. So I will check out for a bit, I have some stuff to do!