2025/08/26 15:39 AI BUG FINALLY FIXED YAY
I redone the Setup AI file and squeezed out the bug. I still have to replace all the old files with the new one, but that is for later. NOW I wanna learn on more how to approach the Enemy AI, for atm I added a bit more of variables which I have no clue what are doing but the exists so that is something...

Anyhow I kinda like this, I hope to find some answers for next time. For now I sign out Goodbye  o/