2025/08/28 19:00 continued
So my Idea is the following:
-Adding Gambits which are in basics a list of instructions that the AI will go through and decide what is the best choice and favors which are basically if the player do good than enemy will less likely to attack that specific player.
-Depending on enemy Personality type it will improve or deprive specific variables. Which makes the enemy into a specific group of acting.
-Attacks will now also include if it is atk ,buff or defense (I will add it into the next update)

I haven't gone over the why I think this is necessary, so here we go!
So my idea is that I think the best approach (for the time and at the time) is personality driven Enemy AI. The theory I want to test is that I will give them aggression and kindness values and make sure that every action will result in a more emotional feel for the enemy, instead the best attack, they will "choose" the one that is the most emotionally the best.

For example if the Player 1 attacked lets say Enemy 2 and Player 2 Lets's say tried to spare it, then Enemy 2 would be more likely to target Player 1 than Player 2 because how Player 1 and 2 interacted with the enemy.

The Goal of this Personality Driven Enemy AI is to simulate cherry picking from multiple players and choosing the worst of them, this could result into the player strategically choosing who will attack the enemy and who tries to spare it, or it could also help if the player 1 way or another.

For another example if lets say the only tries to spare or only tires to kill, then over time it will be either more easier to spare, or harder to kill because of surviving skills. However if for some reason lets say the player does whatever it feels like, than the enemy should respond according to the teams action.

And while yes no player in the right mind would fight and spare at the same time, it could be interesting to see how different enemies would react to different player behavior. So the Absolute goal is:

To make the simple (none boss) fights into fun what ifs, and tests that the players could experiment in. And I could perhaps depending on the species of the enemy I could introduce maybe a way of increasing the difficulty depending if the player tires to attack specific enemy species. Could be cool! The other thing I need to keep in mind that this idea is still in progress and I haven't mapped out a 100% of all possibilities but right now as I said in the beginning I have little idea what I'm doing, but I want to do something either way. I don't want the AI to be as simple as it gets. I want to put a fun twist, my own twist that I haven't seen before, it will be shitty, but I will try to improve it and do it better the second I mess up somewhere.

Alright that is the explanation, now I will just go and sleep myself out GN!