2025/08/29 15:51 🎊 🎊 🎊  Happy 1 Year Anniversary 🎊 🎊 🎊
1 Year passed since I started this project, it is... a weird feeling, I have done the best to stay alive and do my absolute best to create this game for the future and it is unbelievable to say that I've been doing this for a year now, and it is a huge relief to work on this after so many years of wanting to make this game, and after 1 whole year a 365 day of work I can say... I'm not even half way done with the combat XD.

It is one thing to say I been working on this and it's been fun, and it's another to say I'm not even half way done XD. It is unreal to pass a year and feel like it means nothing, but it is also weird to say it does. I think the problem is that I been doing it as passion project for the whole year getting some work done while doing other stuff like work, school, events and just regular life. But to say this is a "HUGE milestone" and "I've CHANGED everything" is a bit on the other side of the spectrum I think I do changed and I do changed a lot of things, but I think I'm a perfectly steady place to call myself on the road and on the journey to finish this and a lot of hardships are still on the road just ahead and on sight. There will be many more challenges and road blocks, both bugs and personal reasons alike. I'm happy to be here and admire the journey so far, even tho I had NO idea how I even want to imagine this game and now  I have more of a game than a starter project, so much I done and so many things need to be done.

If by chance you read this or you were able to read any of my other descriptions I'm thankful, and I hope this will be turned into the thing I originally intended to be... I don't know right now who writes these words what will be or how I will end up... but... I hope at least a piece of my soul and heart could stick into this project and end up in someone else's hand and cherish it! Forever! Thank you and thank you for reading. I hope many more year will be spend staring at my monitor not knowing what tha fuck to even write or type.

and I hope we will hope I can actually finish it : ) ❤️