2025/09/10 18:32 back to the grind
SOOO... it has been 2 weeks! why? Cause I moved to a Collage and had a bit of a rough time doing anything for the past 2 weeks, but I'm back and I will try to do more commits in the future, but I we will see that. For now I want to focus on more of this project and do Collage as a side try to hone my skills and have what it takes to finish this game.

BUT it is kinda hard especially this first week -_-

I stuck into the elevator twice, had to run up and down stairs (to the first floor and back to the 10th), had to get small pictures to my card to get in cause I messed up, I broke my headphones, I broke my scooter (a new one), got oil everywhere and I just can't stop play the GOD DAMN SILKSONG!

So the TL:DR is that I had a rough start and I do want to focus on my project and I have some great roommates (but it is hard to focus, but I will figure it out). But the goal for the future is to finish the game where  I'm happy with the AI. I will get used to Collage and everything but I will need time and right now that is the hardest thing cause of everything, but I hope this isn't a short term plan but a long term one, cause if I can get past the beginning stupidity, than I will be in a much better place where I'm the one in control of my life and project. For now this is an embarrassingly small update but for a specific individual I cannot work anymore today... Thanks Márk...

Anyways tomorrow I will get up tomorrow and I will do everything I wanted today, so I'm not worried. GN and see ya next update