2025/09/17 14:55 Enemy Can now DIE...!
Alright I had fun with this, I could have a 100% done this with a particle node, but I fucked around in Pixel Composer for an hour for this at least 5/10 sprite, and honestly it was worth it, it looks good and I like it. It is another question if I want to get reminded of the fact that I want to implement all the sizes and colors each like this way, but I will cross the way when I get to it, for now it is alright! (also the fact that in theory the player gets the same size and colored flame is bad... but again later)

For now I'm happy with this, and I will get to do some stuff tomorrow too. I had fun, that is it for today, or more if I have more ideas, but yeah, GOOD ByeEEE