2025/09/18 14:01 I did something for sure!
I done some out of energy stuff, it is rough and I don't feel it works with the style of the game we will see if I want to keep it or not. I compressed a bit of the code, but that is it, nothing huge, nothing less. I really should rave up the amount of code I do cause this is sad, the amount of time I have in my hand is a lot and the amount of time I use up for coding is shameful, I will write try to get ADHD medication so maybe that could help me out focusing on both in the lessons and on my project. It is nice to do stuff from my own, but I now more then ever need a bit of a boost to do what I started here, and would love to get an edge on the matter. Anyhow and rent being over I will go now, once again hoping to do some more tonight if not well Silksong got me again.