2025/09/21 20:21 Finished the Reworked Bar System
Oh hey would you look at that! I was able to squeeze the afternoon of today to finish the Bar system, I had a lot of fun with it. Despite its still being buggy and messy at the edges I like it, and I cannot wait for me to start bug hunting. Some of these little critters start get on my nerves and I could spend a day or two on just squishing them, but what is the point if after I'm finished there will be 20 more that will pop up in the course of the project. For now I will have to just ignore them, but I swear as soon as I am finished with the main mechanics I will buy a flamethrower.

Anyways, I added a graphics for Adrenalin for later down the line, and funned around the game to get a feel, and it is so much better then the one I used before.

I do not have to do basically anything I just have to call a premade function that does everything I wanted to do. So that is convenient.

But for real this time I do not think I will have anymore free afternoons so see ya at Wednesday!

o/