2025/09/26 Comments and the end of Mid phase
308...
308 days have passed since I started this phase, and no its been over.

for compressing I've been working on this game for... ALMOST 400 DAYS WHAT??? (393 as of today wow)

so I spent a 100 days to get the basics done. (not exactly but you get it)
And then I spent 3x as much on the mid section of the game (minus like ~60 days where I haven't done anything and the many days where I just didn't do anything)

It is mixed how I feel. I knew a year is a lot of time, but seeing that is been 400 days almost since the first commit, is just wow. I'm so happy I can finally say that we are half way there...
...
...
wow I am just only half way there... oh...fuck...

*spins and shakes head really fast*
OKAY I'm over it, back for whats to come!
I been organizing the trello page and mostly this phase, The END GAME WOHOOO (god damn that is funny) phase is just mostly a revisits of old bad concepts, like Movement, Camera, dialogue, skills and others

The goal still stands, I want to finish the game by the end of THIS year, I know I said that last year, but I'm serious right now! <- serous "!"

I want to finish this with everything I have left to do. So around either in the summer or next fall I have a prototype which I can give out to my friends and I can start with... stage 2... god I'm so far behind.

The other goal of mine is that I want to learn both music and shaders next year so that is gonna take up a lot of time, plus the next semester, plus the fuck knows what else.
*claps hands together*
SO I want to finish most of this As Soon As Possible. I will start recording a video this weekend and I will slowly edit it through out of early October, meanwhile I will work on the movement (blaugh)  again, and I hope I can figure it out by at least the time period.. which would be at least half optimal, but would be nice.

I do not want to start into anything else until I'm not done with this project, I would love to, but if I could just finish it, than I could show it to people and after that I want to meet other people and get some of them on the project. But first I need to create this so I have something to show off.

Back in 2024/11/22 I started to enjoy life, stuff just happened and I was a part of a lot. And I want to do those stuff again too (the now stuff like parties), but for that I need to scratch my itch for that feeling of finishment I want so badly. (I know finishment is not a real world, but fuck it it is now, also it sounds awesome)

So for the time being, I just need to bite the bullet and do the background of the game, and just deal with the fact that I cannot bring that many people to the game creation until I bring this to the table. Because no matter what I say or do, THIS is the start of LOTTH, and I need to do THIS, before anything else, and just deal with any hardships later. I can beg or speak how highly I want to, but until I don't have something to give, I cannot promise, and until I cannot promise I cannot ask, even if I would love to find interesting and creative people to help me out... but if anyone else reading this just thank you for sticking with me, or my creation... I just don't think anyone else would read through these logs if you don't have some connection with me... wow it could sound really dumb if I'm the one reading out loud these, imagine that XD (what an asshole for speaking about his breakfast from like a hundred days ago wow narcissism)

but thank you... even if I'm stuck in the past and hoping to write to a future audience, let that be friends, family or a community I so cherish to create I just love the now and I wish to create the future I see in front of me. <3