2025/10/05 20:48 Alright break is done Back to work!
So it has been 9 whole days since last update. I've worked on the next update videos and other stuff, such as ✨ University stuff ✨ .

But I'm back and someone gave me a great idea how to do a "better" (for now in quotation marks) movement system. The short and of the idea is that I could move inside the given range and I could have a confirm button to advance to the next option or the next player. I like this idea mainly because I do not have to program in a calculation to calculate the ellipses and pin point what was the problem anymore... also I think I know the problem (which basically boils down to I was stupid and the combination of fuck mouse interaction in Godot cause it would be too simple if it could be that simple).

Anyhow I'm done for the day and for probably until Wednesday, cause of more Monday and Thursday stuff, but I will be back and with more changes then ever <3