2025/10/17 15:26 Once again I'm on the train and doing my coding stuff :P
I finished movement for now, I still need to some bug fixing later so it works, but that is for a day when I actually care about bugs and me being delusional. But other then that I also started to work on Enemy movement... Before I even start to program into all the stuff I need to figure out how much one player movement average in Enemy steps, and for right now I would rather play than bang my head against the wall... okay I SHOULD finish it, but... there is always later right? yeah right. So I will figure it out later.

Other then all of this I also added the option to dynamically change the movement circle... again, now it works like this:
a = one unit of circle size
b = one unit of movement range (in human cm)
c = one unit of speed addition
d = half of a (a/2)

Vector2( (a*b) + ((a/2)*c) , (d*b) + ((d/2)*c) )

For none nerd language, we have two dimensions x and y and basically I multiply it by every b which will give me the size of the circle, so...:
1 unit of circle * 15 unit of movement size = 15 size

than I just add speed which is half as effective as movement size, so...:
1/2 unit of circle * 2 unit of movement size = 1 unit of movement size

and then add the two together, 15 + 1 = 16 movement size

that is it, that is the longest explanation for a 1+1 you will ever see!
(I didn't use the actual numbers for the example, but that is regardless)

But anyways I will finish it tomorrow or when I have another 2 and a half hours of nothing but no bs. BYE